refactor: Make helper for form
